+++ src/qx/utility/qx-macros.h
@@ -2,7 +2,7 @@
 #define QX_MACROS_H
 
 #define QX_SCOPED_ENUM_HASH_FUNC(T) \
-inline uint qHash(const T& t, uint seed) { \
+inline size_t qHash(const T& t, size_t seed) { \
     return ::qHash(static_cast<typename std::underlying_type<T>::type>(t), seed); \
 }
